Fix bug in buffer_path_identifier
Rvm's path-identifier API was returning both the complete path along with the
actual ruby being used (what we're interested in for the purpose of
buffer_path_identifier) for files that are within any of the sub-directories of
root. As a fix, I now split the string returned by rvm and return the last
value, which is the correct ruby string.